Reduce font size for embedded forms
Reduces the base font size by the same factor as Sirius themes so that the text embedded forms isn't larger than the rest of the page.

This has to be applied to the `<html>` element so that it's properly cascaded to all other elements.

Also introduces a proper cache-busting query string based on the sha sum of the contents of `./web/static`

Fixes VEGA-1278 and VEGA-1276 #minor