Prevent users from rerunning report
Because they're not given immediate feedback, we believe some users are clicking on the "Download" button multiple times and re-running the report.

This change disables the button after the first time it's pressed, and adds a notification banner to explain that it will open in a new window when ready.

Fixes VEGA-1506 #minor