Sort drafts from newest to oldest
When loading the "edit document" form after creating a draft, the first draft in the list is automatically loaded. It should be sorted from newest to oldest to ensure the just-created draft is shown first.

I've had to sort by ID rather than date because `CreatedDate` is a poorly-formatted string that doesn't sort nicely. Getting a proper date would involve API changes, which seem unnecessarily risky for this functionality.

Fixes VEGA-1759 #patch