Show inserts when selecting document template (#295)
* Show inserts when selecting document template

Adds a component to the template selection screen which shows the available inserts when you select a document. It has the same design and functionality of the existing insert screen but is inline.

I had to create a new data object in a structure that's useful for JavaScript. The component then generates the DOM either from templates or from directly creating each element (using a helper function called createElement).

The component contains a hidden input called `hasViewedInserts` which causes the next page to be skipped. But if the component fails to load then the input isn't added and the user is taken to the fallback page.

Fixes VEGA-1751 #minor

* Sort drafts from newest to oldest

When loading the "edit document" form after creating a draft, the first draft in the list is automatically loaded. It should be sorted from newest to oldest to ensure the just-created draft is shown first.

I've had to sort by ID rather than date because `CreatedDate` is a poorly-formatted string that doesn't sort nicely. Getting a proper date would involve API changes, which seem unnecessarily risky for this functionality.

Fixes VEGA-1759 #patch