Switch from pact stub to Wiremock
Built off the proof-of-concept in #338

The only major difference is that I've separated the existing mocks into `common.json` for mocks we're likely to use across multiple tests and `migrated-from-pact.json` for everything else.

The mocks in `migrated-from-pact.json` should either be tidied up into the common set, or moved closer to their associated test.

#minor