Use go-playground/form
Rather than manually building a struct based on the data sent, describe it with struct tags.

We can avoid defining the `form:X` tag if the form element name matches the variable we're setting it to (as seen with `sirius.Address`).

I also added a `addDefaultCountry` helper function to simplify the creation of the `sirius.Draft` object.

For VEGA-1892 #patch