Scope flash cookie to / path
Without the Path property, the cookie is not correctly matched
when attempting to remove it. This results in the flash message
persisting across requests and appearing incorrectly on different
cases etc.

It looks like this can't be tested against the app running locally
in standalone mode, as the paths appear to be set correctly
and the flash cookie is deleted as expected.