Add help target to Makefile
When you run `make` without a specific target, it runs the first command in the Makefile. CodeQL does this, and it causes the verification to take ~15mins since the current first target (`all`) builds everything and runs the end-to-end tests.

This PR adds a `help` target at the top of the file so that that is what's run by CodeQL.

#patch