[WIP] Check support for govuk-frontend v5
Testing with the changes to govuk-frontend in version 5 of the Design System, alongside related changes in MOJ Pattern Library.

NB: to test locally, you must copy the contents of the MOJPL `package` folder from [its WIP branch](https://github.com/ministryofjustice/moj-frontend/pull/528) into a folder in the root of this repo called `./moj-frontend`.

Changes we have to make:
 - Update path of dist files in govuk-frontend
 - Remove `data-module="govuk-details"` (it no longer does anything)
 - Use `[disabled]` attribute rather than `govuk-button--disabled` class (which has been removed)
 - Update body class initialisation

I haven't added `type="module"` to the `<script>` tag because we don't _need_ to, but we might like to do so for closer alignment with GDS guidance.

#major