Use a less demanding warning type
Using "Donor Deceased" as the warning causes issues, as every
case for the donor has to be associated with this warning.

By using a different warning type, we can just attach it to a
single case.

This makes a difference when writing the provider tests as it
simplifies them: we can make one case and one warning per pact, and
don't have to concern ourselves with attaching each warning we
create to all of the cases for the donor.