Mounting a filesystem onto a container removes it from running one
Running `yarn watch` in a docker container and mounting the
directory already mounted into the running lpa-frontend container
wipes out the files in that mount. They are still there, but
the second mount appears to unmount the directory from the
existing container.

For now, remove the yarn watch command while I sort it out.