Providing we do yarn build first, everything is fine
It looks like this might be an issue with directories not existing
at the point when the mount is established.

If we build all the assets first, then mount the directory into
lpa-frontend, it appears that we are able to watch and refresh
the JS file, and it gets picked up correctly.