Show LPA details if in store
If the LPA's status is "Draft" but there's data in the store, show the data in the store.

This is initiated by making testing easier, but is also part of a move towards not relying on Sirius to determine a case's status.

Also, fix the whitespace around `newlines-as-br` things: excess whitespace (newlines/indent) was being rendered inappropriately, replace it with a `<p>`.

Fixes VEGA-2578 #minor