Handle validation errors from editing documents (#576)
Publishing a document involves four API requests. Currently any error from these is treated as a system error and handled like a 500, but a validation error should be handled like a 400 and passed into the template.

To avoid doing this four times, I've moved the API calls into a single serial function. If any error is returned, the outer code checks if it's a validation error and handles it more carefully if so.

#patch