Fix accessibility issues in search
- Add `required` to the search input to prevent submission with enter
- Remove unnecessary `aria=label` (it already has a real label)
- Update label to match placeholder

This stops two things on the page (the search box and the search button) having the same label text ("Search"), which will hopefully make the page easier to navigate with voice control.

My expectation is that voice control users should be able to say "Click Search for a case" to get to the input (matching what they can see on screen). But they could also say "Click Search" and it will try to submit the form, fail because the required field is empty, and focus on the input field.

This is hard to test, because each operating system/browser/tool handles things slightly differently, but I've had relatively positive results on Voice Control for Mac.

#patch