Switch to Electron for Cypress tests and fix (#791)
Electron is the only [reliably included browser](https://github.com/cypress-io/cypress-docker-images/blob/master/included/README.md), ensuring that our Cypress tests can be run on any machines, and both ARM and AMD infrastructure.

However, it balks when trying to preview a document because the content is empty. To resolve this, use `sirius-mock` as the public Sirius URL and intercept-and-stub the download request.

#patch