Override max length of govuk tags to fix styling
The tag background colour set by govuk-tag has a max length so the
colour cuts off towards the end of our long statuses. This sets an
override class for the css.

There is a generic search tag template which I replaced because the
style would not be suitable for other uses of the status tags such as
the tabs on the LPA case details page